+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- [hardis:doc:project2markdown](https://sfdx-hardis.cloudity.com/hardis/doc/project2markdown/) Generate Apex classes documentation using `@cparra/apexdocs`, and describe them using AI
 - Display error message in case of XML parsing error
 - Do not raise issues when managed items fields don't have descriptions
 - Do not raise inactive validation rule issue when the VR is from aa managed package